# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/execute-request": {
      "post": {
        "description": "Makes an HTTP request to the specified URL with provided configuration and returns the response",
        "operationId": "AppController_executeRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Request executed successfully"
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error or request execution failed"
          }
        },
        "summary": "Execute an HTTP request with given details",
        "tags": [
          "App"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS and Express API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "ExecuteRequestDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "The URL to send the request to"
          },
          "method": {
            "type": "string",
            "enum": [
              "GET",
              "POST",
              "PUT",
              "DELETE",
              "PATCH",
              "HEAD",
              "OPTIONS"
            ],
            "default": "GET"
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "data": {
            "type": "object",
            "description": "Request body data"
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "responseType": {
            "type": "string",
            "enum": [
              "json",
              "text",
              "stream",
              "blob",
              "document",
              "arraybuffer"
            ],
            "default": "json"
          },
          "timeout": {
            "type": "number",
            "description": "Request timeout in milliseconds",
            "default": 30000,
            "minimum": 1000,
            "maximum": 300000
          },
          "followRedirects": {
            "type": "boolean",
            "description": "Whether to follow redirects",
            "default": true
          },
          "maxRedirects": {
            "type": "number",
            "description": "Maximum number of redirects to follow",
            "default": 5,
            "minimum": 0,
            "maximum": 10
          }
        },
        "required": [
          "url"
        ]
      },
...
    }
  }
}